rf: move `compile_dandiset_validation_report`
From `cli/tools.py` to `tools.py`
